requirements.txt
SpeechRecognition
pipwin
pyaudio
gTTS
google-generativeai
python-dotenv
streamlit

-e .

